refactor(ops): move KDA torch adapters into operator directories
