+++ csrc/torch_binding.cpp
@@ -45,6 +45,9 @@
 #include "moe/causal_conv1d_v310/causal_conv1d_310_torch_adpt.h"
 #include "attention/recurrent_gated_delta_rule/recurrent_gated_delta_rule_torch_adpt.h"
 #include "attention/recurrent_kda/recurrent_kda_torch_adpt.h"
+#include "attention/chunk_kda_fwd/chunk_kda_fwd_torch_adpt.h"
+#include "attention/kda_gate_cumsum/kda_gate_cumsum_torch_adpt.h"
+#include "attention/kda_layout_swap12/kda_layout_swap12_torch_adpt.h"
 #include "attention/recurrent_gated_delta_rule_v310/recurrent_gated_delta_rule_310_torch_adpt.h"
 #include "attention/k2q_csr/k2q_csr_torch_adpt.h"
 #include "attention/msa_index_score/msa_index_score_torch_adpt.h"
@@ -1921,320 +1924,6 @@ at::Tensor npu_sparse_attention_score_prefill(
     );
 
     return output;
-}
-
-namespace {
-
-int64_t kda_ceil_div(int64_t x, int64_t y)
-{
-    return (x + y - 1) / y;
-}
-
-int64_t get_kda_seq_num(int64_t batch, const c10::optional<at::IntArrayRef> &cu_seqlens)
-{
-    if (!cu_seqlens.has_value()) {
-        return batch;
-    }
-    return static_cast<int64_t>(cu_seqlens.value().size()) - 1;
-}
-
-void check_kda_cu_seqlens(const c10::optional<at::IntArrayRef> &cu_seqlens,
-                          int64_t total_tokens,
-                          const char *op_name)
-{
-    if (!cu_seqlens.has_value()) {
-        return;
-    }
-    auto cu = cu_seqlens.value();
-    TORCH_CHECK(cu.size() >= 2, op_name, ": cu_seqlens must contain at least [0, total_tokens].");
-    TORCH_CHECK(cu[0] == 0, op_name, ": cu_seqlens[0] must be 0, but got ", cu[0], ".");
-    TORCH_CHECK(cu[cu.size() - 1] == total_tokens,
-                op_name, ": cu_seqlens[-1] must equal sequence length ",
-                total_tokens, ", but got ", cu[cu.size() - 1], ".");
-    for (size_t i = 0; i + 1 < cu.size(); ++i) {
-        TORCH_CHECK(cu[i] <= cu[i + 1],
-                    op_name, ": cu_seqlens must be nondecreasing, but cu_seqlens[",
-                    i, "]=", cu[i], " > cu_seqlens[", i + 1, "]=", cu[i + 1], ".");
-    }
-}
-
-void check_kda_chunk_indices(const c10::optional<at::IntArrayRef> &chunk_indices,
-                             const c10::optional<at::IntArrayRef> &cu_seqlens,
-                             int64_t chunk_size,
-                             const char *op_name)
-{
-    if (!chunk_indices.has_value()) {
-        return;
-    }
-    auto indices = chunk_indices.value();
-    TORCH_CHECK(indices.size() % 2 == 0,
-                op_name, ": chunk_indices must contain (seq_id, chunk_id) pairs, but got ",
-                indices.size(), " elements.");
-    TORCH_CHECK(cu_seqlens.has_value(), op_name, ": chunk_indices requires cu_seqlens.");
-    auto cu = cu_seqlens.value();
-    int64_t expected_chunks = 0;
-    for (size_t seq = 0; seq + 1 < cu.size(); ++seq) {
-        expected_chunks += kda_ceil_div(cu[seq + 1] - cu[seq], chunk_size);
-    }
-    TORCH_CHECK(static_cast<int64_t>(indices.size() / 2) == expected_chunks,
-                op_name, ": chunk_indices must contain exactly one pair per chunk.");
-    for (size_t idx = 0; idx < indices.size(); idx += 2) {
-        int64_t seq = indices[idx];
-        int64_t chunk = indices[idx + 1];
-        TORCH_CHECK(seq >= 0 && seq + 1 < static_cast<int64_t>(cu.size()),
-                    op_name, ": chunk_indices seq_id is out of range.");
-        int64_t chunks = kda_ceil_div(cu[seq + 1] - cu[seq], chunk_size);
-        TORCH_CHECK(chunk >= 0 && chunk < chunks,
-                    op_name, ": chunk_indices chunk_id is out of range.");
-    }
-}
-
-int64_t get_kda_total_chunks(int64_t batch,
-                             int64_t seqlen,
-                             int64_t chunk_size,
-                             const c10::optional<at::IntArrayRef> &cu_seqlens,
-                             const c10::optional<at::IntArrayRef> &chunk_indices)
-{
-    if (chunk_indices.has_value()) {
-        return static_cast<int64_t>(chunk_indices.value().size()) / 2;
-    }
-    if (!cu_seqlens.has_value()) {
-        return kda_ceil_div(seqlen, chunk_size);
-    }
-    (void)batch;
-    int64_t total = 0;
-    auto cu = cu_seqlens.value();
-    for (size_t i = 0; i + 1 < cu.size(); ++i) {
-        total += kda_ceil_div(cu[i + 1] - cu[i], chunk_size);
-    }
-    return total;
-}
-
-std::vector<int64_t> build_kda_chunk_indices(at::IntArrayRef cu_seqlens, int64_t chunk_size)
-{
-    std::vector<int64_t> indices;
-    int64_t total_chunks = 0;
-    for (size_t i = 0; i + 1 < cu_seqlens.size(); ++i) {
-        total_chunks += kda_ceil_div(cu_seqlens[i + 1] - cu_seqlens[i], chunk_size);
-    }
-    indices.reserve(static_cast<size_t>(total_chunks * 2));
-    for (size_t seq = 0; seq + 1 < cu_seqlens.size(); ++seq) {
-        int64_t seq_len = cu_seqlens[seq + 1] - cu_seqlens[seq];
-        int64_t chunks = kda_ceil_div(seq_len, chunk_size);
-        for (int64_t chunk = 0; chunk < chunks; ++chunk) {
-            indices.push_back(static_cast<int64_t>(seq));
-            indices.push_back(chunk);
-        }
-    }
-    return indices;
-}
-
-} // namespace
-
-std::tuple<at::Tensor, at::Tensor, at::Tensor, at::Tensor, at::Tensor, at::Tensor,
-           at::Tensor, at::Tensor, at::Tensor, at::Tensor, at::Tensor, at::Tensor>
-chunk_kda_fwd(
-    const at::Tensor &q,
-    const at::Tensor &k,
-    const at::Tensor &v,
-    const at::Tensor &gk,
-    const at::Tensor &beta,
-    double scale,
-    int64_t chunk_size,
-    c10::string_view layout,
-    const c10::optional<at::Tensor> &initial_state,
-    c10::optional<bool> output_final_state,
-    c10::optional<at::IntArrayRef> cu_seqlens,
-    c10::optional<at::IntArrayRef> chunk_indices,
-    c10::optional<bool> return_intermediate,
-    c10::optional<bool> safe_gate,
-    c10::optional<bool> transpose_state_layout)
-{
-    std::string layout_str(layout.data(), layout.size());
-    TORCH_CHECK(layout_str == "BSND" || layout_str == "BNSD" || layout_str == "TND" || layout_str == "NTD",
-                "chunk_kda_fwd: layout must be one of BSND, BNSD, TND, NTD and must be uppercase.");
-    TORCH_CHECK(!safe_gate.value_or(false), "chunk_kda_fwd: safe_gate=True is not supported.");
-    TORCH_CHECK(!transpose_state_layout.value_or(false),
-                "chunk_kda_fwd: transpose_state_layout=True is not supported.");
-    TORCH_CHECK(chunk_size == 32 || chunk_size == 64 || chunk_size == 128,
-                "chunk_kda_fwd: chunk_size must be 32, 64 or 128.");
-
-    bool is_tnd = layout_str == "TND";
-    bool is_ntd = layout_str == "NTD";
-    bool is_bsnd = layout_str == "BSND";
-    bool is_bnsd = layout_str == "BNSD";
-    bool is_rank3 = is_tnd || is_ntd;
-    TORCH_CHECK((is_rank3 && q.dim() == 3 && k.dim() == 3 && v.dim() == 3 && gk.dim() == 3 && beta.dim() == 2) ||
-                    (!is_rank3 && q.dim() == 4 && k.dim() == 4 && v.dim() == 4 && gk.dim() == 4 && beta.dim() == 3),
-                "chunk_kda_fwd: layout/rank mismatch.");
-    TORCH_CHECK(q.sizes() == k.sizes(), "chunk_kda_fwd: q and k must have identical shape.");
-
-    auto q_sizes = q.sizes();
-    auto v_sizes = v.sizes();
-    bool is_internal_layout = is_bnsd || is_ntd;
-    int64_t B = is_rank3 ? 1 : q_sizes[0];
-    int64_t T = is_tnd ? q_sizes[0] : (is_ntd ? q_sizes[1] : (is_bnsd ? q_sizes[2] : q_sizes[1]));
-    int64_t H = is_tnd ? q_sizes[1] : (is_ntd ? q_sizes[0] : (is_bnsd ? q_sizes[1] : q_sizes[2]));
-    int64_t K = is_rank3 ? q_sizes[2] : q_sizes[3];
-    int64_t HV = is_tnd ? v_sizes[1] : (is_ntd ? v_sizes[0] : (is_bnsd ? v_sizes[1] : v_sizes[2]));
-    int64_t V = is_rank3 ? v_sizes[2] : v_sizes[3];
-    TORCH_CHECK(H > 0 && HV >= H, "chunk_kda_fwd: H and HV must be positive and H must be <= HV.");
-    TORCH_CHECK(H <= 128 && HV <= 128, "chunk_kda_fwd: H and HV must be <= 128.");
-    TORCH_CHECK(!is_tnd || H == 1,
-                "chunk_kda_fwd: TND layout with H > 1 is not supported; use NTD for multi-head rank3 input.");
-    check_kda_cu_seqlens(cu_seqlens, T, "chunk_kda_fwd");
-    check_kda_chunk_indices(chunk_indices, cu_seqlens, chunk_size, "chunk_kda_fwd");
-    TORCH_CHECK(!cu_seqlens.has_value() || is_rank3 || B == 1,
-                "chunk_kda_fwd: rank4 varlen input with cu_seqlens currently requires B=1.");
-    TORCH_CHECK(HV % H == 0, "chunk_kda_fwd: HV must be divisible by H.");
-    TORCH_CHECK(q.scalar_type() == at::kHalf || q.scalar_type() == at::kBFloat16,
-                "chunk_kda_fwd: q/k/v must use float16 or bfloat16.");
-    TORCH_CHECK(k.scalar_type() == q.scalar_type() && v.scalar_type() == q.scalar_type(),
-                "chunk_kda_fwd: q/k/v dtype must match.");
-    TORCH_CHECK(chunk_size == 64 && K >= 16 && V >= 16 && K % 16 == 0 && V % 16 == 0 && V <= 256 &&
-                    K * V >= 4 * 64 * 64 && K * V >= chunk_size * (K + V),
-                "chunk_kda_fwd: shape is outside the supported split cube/vector template.");
-
-    int64_t seq_num = get_kda_seq_num(B, cu_seqlens);
-    at::Tensor initial_state_tensor = initial_state.value_or(at::Tensor());
-    if (initial_state_tensor.defined()) {
-        TORCH_CHECK(initial_state_tensor.scalar_type() == at::kFloat,
-                    "chunk_kda_fwd: initial_state must be float32 when provided.");
-        TORCH_CHECK(initial_state_tensor.dim() == 4 && initial_state_tensor.size(0) == seq_num &&
-                        initial_state_tensor.size(1) == HV && initial_state_tensor.size(2) == K &&
-                        initial_state_tensor.size(3) == V,
-                    "chunk_kda_fwd: initial_state must be [seq_num,Hv,K,V].");
-    }
-
-    std::vector<int64_t> generated_chunk_indices;
-    c10::optional<at::IntArrayRef> chunk_indices_for_call;
-    if (chunk_indices.has_value()) {
-        chunk_indices_for_call = chunk_indices.value();
-    } else if (cu_seqlens.has_value()) {
-        generated_chunk_indices = build_kda_chunk_indices(cu_seqlens.value(), chunk_size);
-        chunk_indices_for_call = at::IntArrayRef(generated_chunk_indices);
-    } else {
-        chunk_indices_for_call = c10::nullopt;
-    }
-
-    int64_t total_chunks = get_kda_total_chunks(B, T, chunk_size, cu_seqlens, chunk_indices_for_call);
-    at::Tensor o = at::empty_like(v);
-    at::Tensor final_state_work = at::empty({seq_num, HV, K, V}, q.options().dtype(at::kFloat));
-    at::Tensor aqk = is_rank3 ? (is_internal_layout ? at::empty({HV, T, chunk_size}, q.options()) :
-        at::empty({T, HV, chunk_size}, q.options())) : (is_internal_layout ?
-        at::empty({B, HV, T, chunk_size}, q.options()) : at::empty({B, T, HV, chunk_size}, q.options()));
-    at::Tensor akk = at::empty_like(aqk);
-    at::Tensor w = is_rank3 ? (is_internal_layout ? at::empty({HV, T, K}, q.options()) :
-        at::empty({T, HV, K}, q.options())) : (is_internal_layout ?
-        at::empty({B, HV, T, K}, q.options()) : at::empty({B, T, HV, K}, q.options()));
-    at::Tensor u = at::empty_like(v);
-    at::Tensor qg = at::empty_like(w);
-    at::Tensor kg = at::empty_like(w);
-    at::Tensor v_new = at::empty_like(v);
-    at::Tensor h = is_rank3 ? (is_internal_layout ? at::empty({HV, total_chunks, K, V}, q.options()) :
-        at::empty({total_chunks, HV, K, V}, q.options())) : (is_internal_layout ?
-        at::empty({B, HV, total_chunks, K, V}, q.options()) :
-        at::empty({B, total_chunks, HV, K, V}, q.options()));
-
-    bool recompute_output_final_state = true;
-    char *layout_cstr = const_cast<char *>(layout_str.c_str());
-    EXEC_NPU_CMD(
-        aclnnChunkKdaFwd,
-        q, k, v, gk, beta, initial_state_tensor,
-        cu_seqlens, chunk_indices_for_call,
-        layout_cstr, scale, chunk_size, recompute_output_final_state, total_chunks,
-        o, final_state_work, aqk, akk, w, u, qg, kg, v_new, h
-    );
-
-    at::Tensor final_state = output_final_state.value_or(false) ?
-        final_state_work : at::empty({0}, q.options().dtype(at::kFloat));
-    at::Tensor empty = at::empty({0}, q.options());
-    at::Tensor g = gk.scalar_type() == at::kFloat ? gk : gk.to(at::kFloat);
-    at::Tensor initial_state_out = initial_state_tensor.defined() ? initial_state_tensor : empty;
-    (void)return_intermediate;
-    return std::make_tuple(o, final_state, g, aqk, akk, w, u, qg, kg, v_new, h, initial_state_out);
-}
-
-at::Tensor kda_gate_cumsum(
-    const at::Tensor &g,
-    int64_t chunk_size,
-    const c10::optional<at::Tensor> &A_log,
-    const c10::optional<at::Tensor> &dt_bias,
-    c10::optional<at::IntArrayRef> cu_seqlens,
-    c10::optional<bool> use_gate_in_kernel,
-    c10::optional<bool> safe_gate,
-    c10::optional<double> lower_bound,
-    c10::string_view layout)
-{
-    TORCH_CHECK(g.dim() == 3 || g.dim() == 4,
-                "kda_gate_cumsum: g must be BSND/BNSD rank4 or TND/NTD rank3.");
-    TORCH_CHECK(chunk_size == 32 || chunk_size == 64 || chunk_size == 128,
-                "kda_gate_cumsum: chunk_size must be 32, 64 or 128.");
-    auto gate_dtype = g.scalar_type();
-    TORCH_CHECK(gate_dtype == at::kFloat || gate_dtype == at::kBFloat16 || gate_dtype == at::kHalf,
-                "kda_gate_cumsum: g must be float32, bfloat16 or float16.");
-    std::string layout_str(layout.data(), layout.size());
-    TORCH_CHECK(layout_str == "BSND" || layout_str == "BNSD" || layout_str == "TND" || layout_str == "NTD",
-                "kda_gate_cumsum: layout must be uppercase and one of BSND, BNSD, TND or NTD.");
-    bool is_bnsd = layout_str == "BNSD";
-    bool is_ntd = layout_str == "NTD";
-    int64_t T = is_bnsd ? g.sizes()[2] : (is_ntd ? g.sizes()[1] : (g.dim() == 4 ? g.sizes()[1] : g.sizes()[0]));
-    int64_t K = g.dim() == 4 ? g.sizes()[3] : g.sizes()[2];
-    int64_t HV = is_bnsd ? g.sizes()[1] : (is_ntd ? g.sizes()[0] : (g.dim() == 4 ? g.sizes()[2] : g.sizes()[1]));
-    TORCH_CHECK(K <= 256, "kda_gate_cumsum: K must be <= 256.");
-    check_kda_cu_seqlens(cu_seqlens, T, "kda_gate_cumsum");
-    TORCH_CHECK(!cu_seqlens.has_value() || g.dim() == 3 || g.sizes()[0] == 1,
-                "kda_gate_cumsum: rank4 varlen input with cu_seqlens currently requires B=1.");
-
-    bool use_gate = use_gate_in_kernel.value_or(false);
-    bool safe = safe_gate.value_or(false);
-    double lower = lower_bound.value_or(-5.0);
-    at::Tensor A_log_tensor = A_log.value_or(at::Tensor());
-    at::Tensor dt_bias_tensor = dt_bias.value_or(at::Tensor());
-    if (use_gate) {
-        TORCH_CHECK(A_log_tensor.defined(), "kda_gate_cumsum: A_log is required when use_gate_in_kernel=True.");
-        TORCH_CHECK(A_log_tensor.scalar_type() == at::kFloat && A_log_tensor.dim() == 1 && A_log_tensor.sizes()[0] == HV,
-                    "kda_gate_cumsum: A_log must be float32 with shape [HV].");
-        TORCH_CHECK(safe, "kda_gate_cumsum: raw gate path currently requires safe_gate=True.");
-        TORCH_CHECK(lower >= -5.0 && lower < 0.0, "kda_gate_cumsum: lower_bound must be in [-5, 0).");
-    } else {
-        TORCH_CHECK(!safe, "kda_gate_cumsum: safe_gate only applies when use_gate_in_kernel=True.");
-    }
-
-    at::Tensor gk = at::empty(g.sizes(), g.options().dtype(at::kFloat));
-    char *layout_cstr = const_cast<char *>(layout_str.c_str());
-    EXEC_NPU_CMD(
-        aclnnKdaGateCumsum,
-        g, A_log_tensor, dt_bias_tensor, cu_seqlens,
-        chunk_size, use_gate, safe, lower, layout_cstr, gk
-    );
-    return gk;
-}
-
-at::Tensor kda_layout_swap12(
-    const at::Tensor &x,
-    const c10::optional<at::Tensor> &dependency)
-{
-    TORCH_CHECK(x.dim() >= 3, "kda_layout_swap12: x must have rank >= 3.");
-    auto dtype = x.scalar_type();
-    TORCH_CHECK(dtype == at::kFloat || dtype == at::kHalf || dtype == at::kBFloat16,
-                "kda_layout_swap12: x must be float32, float16 or bfloat16.");
-
-    std::vector<int64_t> y_sizes(x.sizes().begin(), x.sizes().end());
-    if (x.dim() == 3) {
-        std::swap(y_sizes[0], y_sizes[1]);
-    } else {
-        std::swap(y_sizes[1], y_sizes[2]);
-    }
-    at::Tensor y = at::empty(y_sizes, x.options());
-    at::Tensor dependency_tensor = dependency.value_or(at::Tensor());
-    if (dependency_tensor.defined()) {
-        TORCH_CHECK(dependency_tensor.sizes() == y.sizes(),
-                    "kda_layout_swap12: dependency must have the same shape as output.");
-    }
-
-    EXEC_NPU_CMD(aclnnKdaLayoutSwap12, x, dependency_tensor, y);
-    return y;
 }
 
 std::vector<int64_t> get_npu_storage_shape(const at::Tensor& tensor)
